ray_deployment/ray_deployment/requirements.txt
albumentations==1.3.0
comet_ml==3.32.3
coremltools==6.2
fastapi==0.92.0
Flask==2.1.0
GitPython==3.1.31
ipython==8.11.0
matplotlib==3.6.3
mss==7.0.1
numpy==1.24.1
onnx==1.13.1
onnxruntime==1.14.1
onnxsim==0.4.17
opencv_python==4.7.0.68
openvino==2022.3.0
paddle==1.0.2
pafy==0.5.5
pandas==1.5.3
Pillow==9.4.0
psutil==5.9.4
pycocotools==2.0.6
pymongo==4.3.3
PyYAML==6.0
qoa4ml==0.0.63
ray==2.2.0
requests==2.25.1
scipy==1.10.1
seaborn==0.12.2
setuptools==65.5.0
starlette==0.25.0
tensorflow==2.7.0
tensorflowjs==4.2.0
tensorrt==8.5.3.1
tflite_runtime==2.11.0
tflite_support==0.4.3
thop==0.1.1.post2209072238
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
tritonclient==2.31.0
ultralytics==8.0.25
x2paddle==1.4.0
